[Refactor] split mapper package
