+++ app/src/main/java/com/and04/naturealbum/ui/maps/MapScreenViewModel.kt
@@ -3,6 +3,9 @@ package com.and04.naturealbum.ui.maps
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.and04.naturealbum.data.repository.DataRepository
+import com.and04.naturealbum.data.room.Label
+import com.and04.naturealbum.data.room.PhotoDetail
+import com.naver.maps.geometry.LatLng
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.async
 import kotlinx.coroutines.flow.MutableStateFlow
@@ -24,5 +27,19 @@ class MapScreenViewModel @Inject constructor(
 
             _photos.emit(fetchPhotos.await().toPhotoItems(fetchLabels))
         }
+    }
+}
+
+// Room Data -> UI Data
+fun Label.toLabelItem() = LabelItem(name, backgroundColor)
+fun List<PhotoDetail>.toPhotoItems(labels: List<Label>): List<PhotoItem> {
+    val labelMap = labels.associate { roomLabel -> roomLabel.id to roomLabel.toLabelItem() }
+    return map { photoDetail ->
+        PhotoItem(
+            photoDetail.photoUri,
+            LatLng(photoDetail.latitude, photoDetail.longitude),
+            labelMap.getValue(photoDetail.labelId),
+            photoDetail.datetime
+        )
     }
 }
